requirements.txt
semantic_version
requests
beautifulsoup4
pyfpgrowth
mysql-connector-python